Fixed issues
Problem 1: age of first and second persons were not calculating correctly
Solution 1: include parantheses to clarify order of operations. now, it determines whether there is a year of death or not, and then calculates age

Problem 2: people sorted incorrectly
Solution 2: if ageOfFirstPerson is greater than ageOfSecondPerson, value should return -1 to bubble the person up to the front of the list.